em100: Support 4 byte address mode
This change adds a command line option to set 3 or 4 byte address
mode for the em100 FPGA.  It automatically sets 4 byte address mode
if the selected chip is >16MB, unless 3 byte address mode is set
on the command line.

This does not seem to be necessary with the most recent FPGA version
2.014 but for devices with older versions (I have 2.010) this allows
emulating 32MB SPI flash parts like W25Q256.

Change-Id: Ibe955eb3b63bc51ae618d176a16d8a50f55c8c55
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://review.coreboot.org/c/em100/+/40418
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>